Extract suggestion related UI and logic to a separate file
